refactor: ReservationRequestFactory를 서비스 객체로 변경
- 클래스명을 BookService로 수정
- 예약 생성 및 삭제 메서드를 본 클래스로 이동
